+++ tests/infrastructures/Kubernetes/Transcriber/SecretTranscriberTest.php
@@ -60,6 +60,11 @@ class SecretTranscriberTest extends TestCase
             });
 
         $seRepo = $this->createMock(SecretRepository::class);
+
+        $kubeClient->expects(self::atLeastOnce())
+            ->method('setNamespace')
+            ->with('default_namespace');
+
         $kubeClient->expects(self::any())
             ->method('__call')
             ->willReturnMap([